+++ packages/tests-render-mithril-css/src/components/navigation-list.js
@@ -6,7 +6,7 @@ const icons = {
   send: "<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"24\" height=\"24\" viewBox=\"0 0 24 24\"><path d=\"M2.01 21L23 12 2.01 3 2 10l15 2-15 2z\"/></svg>"
 };
 
-// Style for class "tests-drawer-navigation-list" is created in scripts/writeThemCSS.js
+// Style for class "tests-drawer-navigation-list" is created in scripts/writeThemeCSS.js
 
 export default ({ renderer: h, keys: k, Icon, List, ListTile, onClick=()=>{} }) => {